155 commit

+++ app/src/main/java/com/pozmaxpav/cinemaopinion/presentation/components/detailscards/DetailsCardSeasonalEvent.kt
@@ -69,7 +69,7 @@ fun DetailsCard(
     }
 
     LaunchedEffect(newYearMovie.id) {
-        mainViewModel.getInformationMovie(newYearMovie.id)
+        mainViewModel.getInformationMovie(newYearMovie.id.toInt())
     }
 
     Column(
